tags in OZ_Container should be uintptr_t

+++ platform/emulator/mozart.h
@@ -707,6 +707,7 @@ VAR = OZ_vsToC(OZ_in(ARG),&LEN);
 #if defined(__cplusplus)
 
 #include <stdlib.h>
+#include <stdint.h>
 
 /* some internal functions to make the extension class work */
 _FUNDECL(void*,_OZ_new_OZ_Extension,(size_t n));
@@ -717,21 +718,21 @@ _FUNDECL(void*,_OZ_currentBoard,());
 
 class OZ_Container {
 public:
-  unsigned int tag;
-  void init(unsigned int t) {
+  uintptr_t tag;
+  void init(uintptr_t t) {
     tag = t << 1;
   }
   void initAsExtension(void) {
     init(OZ_CONTAINER_TAG);
   }
-  unsigned int cacIsMarked(void) { 
+  uintptr_t cacIsMarked(void) { 
     return tag&1; 
   }
   void cacMark(OZ_Container * c) { 
-    tag = ((unsigned int) c) | 1; 
+    tag = ((uintptr_t) c) | 1; 
   }
-  unsigned int ** cacGetMarkField(void) { 
-    return (unsigned int **) &tag; 
+  uintptr_t ** cacGetMarkField(void) { 
+    return (uintptr_t **) &tag; 
   }
   OZ_Container * cacGetFwd(void) {
     Assert(cacIsMarked());
